+++ HISTORY.md
@@ -7,10 +7,11 @@ are in `bcbioRNASeq/results/date/genes/counts` and
 `bcbioRNASeq/results/data/transcripts/counts`.
 - qualimap (RNA-seq): disable stranded mode for qualimap, as it gives incorrect
 results with the hisat2 aligner and for RNA-seq just setting it to unstranded
-gives reasonable results.
+<<<<<<< HEAD
 - Add `quantify_genome_alignments` option to use genome alignments to quantify
   with Salmon.
 - Add `--validateMappings` flag to Salmon read quantification mode.
+- VEP cache is not installing anymore from bcbio run
 
 ## 1.1.5 (12 April 2019)